requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tavily-python==0.3.0
groq==0.4.2
prometheus-client==0.19.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
